feature/2.2.x-kart-saklama update kart sakalama fix

+++ admin/model/payment/iyzico_checkout_form.php
@@ -3,6 +3,7 @@
 class ModelPaymentIyzicoCheckoutForm extends Model {
 
     public function install() {
+		if(!isset($this->session->data['iyzico_update'])){
         $this->db->query("
 			CREATE TABLE IF NOT EXISTS `" . DB_PREFIX . "iyzico_order` (
 			`iyzico_order_id` INT(11) NOT NULL AUTO_INCREMENT,
@@ -35,10 +36,10 @@ class ModelPaymentIyzicoCheckoutForm extends Model {
 				ADD COLUMN `card_key` VARCHAR(50),
 				ADD COLUMN `iyzico_api` VARCHAR(100),
 				ENGINE=MyISAM DEFAULT COLLATE=utf8_general_ci;");
-		
+		}
 		$file = DIR_APPLICATION . 'iyzico_checkout_form.sql';
 		if (!file_exists($file)) { 
-			}
+		}
 		$lines = file($file);
 		if ($lines) {
 			$sql = '';
@@ -58,11 +59,14 @@ class ModelPaymentIyzicoCheckoutForm extends Model {
     }
 
     public function uninstall() {
+		if(!isset($this->session->data['iyzico_update'])){
         $this->db->query("DROP TABLE IF EXISTS `" . DB_PREFIX . "iyzico_order`;");
         $this->db->query("DROP TABLE IF EXISTS `" . DB_PREFIX . "iyzico_order_refunds`;");
 		$this->db->query("ALTER TABLE `" . DB_PREFIX . "customer` DROP COLUMN card_key;");
 		$this->db->query("ALTER TABLE `" . DB_PREFIX . "customer` DROP COLUMN iyzico_api;");
+		}
 		$this->db->query("DELETE FROM `" . DB_PREFIX . "modification` WHERE code='iyzico_checkout_form'");		
+		
     }
 
     public function logger($message) {
